[ADD] project_mail_plugin: add support for project's tasks in mail plugin
Add support for tasks for the mail plugins via the new project_mail_plugin
module.

The new module provides the following new features:

- Enables the user to see tasks related to the selected contact, tasks where the
  contact is the customer

- Enables the user to log an email directly to a task as this can be useful when
  the user finds important information in the email

- Enable the user to create a task with the email message being directly set
  in the description field, as the user would probable want the email
  message to be included in the created task

Purpose is to allow users to connect their tasks to their mailbox.

Which allows for example to create a task from an email received,centralize data
regarding a customer on its task, check ongoing tasks for a customer, ...etc

Add the fontawesome-4.7 package in dependencies, purpose is to allow usage of
some icons which are not available in v5, especially some icons which are used
in the backend as the backend is based on v4.7

Task-2518640
COM-PR: https://github.com/odoo-dev/odoo/pull/882