[IMP] outlook: redirect to documentation when mail plugin module not found
Before logging the user in, test if the mail plugin module is installed, if it's
not the case, redirect the user to the documentation.

Purpose is to help users understand what went wrong and show them how they can
fix the issue.

COM-PR: https://github.com/odoo/odoo/pull/73281
Task-2577531